refactored VorDiagram. Added DCEL structure
