+++ Makefile
@@ -5,7 +5,8 @@ GraphGenerator: GraphGenerator.class
 
 # Compiling DCEL classes
 DCEL_DIR = DCEL/
-DCEL_SRC = $(DCEL_DIR)HalfEdge.java $(DCEL_DIR)Vertex.java $(DCEL_DIR)Face.java $(DCEL_DIR)HalfEdgeBuilder.java
+DCEL_SRC = $(DCEL_DIR)HalfEdge.java $(DCEL_DIR)Vertex.java $(DCEL_DIR)Face.java $(DCEL_DIR)HalfEdgeBuilder.java \
+					 $(DCEL_DIR)DCEL.java
 DCEL_CLASSES = $(DCEL_SRC:.java=.class)
 $(DCEL_CLASSES): $(DCEL_SRC)
 	javac -g $(DCEL_SRC)
