+++ VoronoiDiagram.java
@@ -12,197 +12,29 @@ import Utils.*;
 import Utils.Exceptions.*;
 import DCEL.*;
 
-public class VoronoiDiagram implements VisualData{
-  private ArrayList<HalfEdge> edge;
-  private ArrayList<Vertex> vertex;
-  private ArrayList<Face> face;
-  private HashMap<Point, Face> face_searcher;
-  
+public class VoronoiDiagram extends DCEL implements VisualData{
+  // Max coordinates of diagram
   public static final double maxX = 1000000.0;
   public static final double minX = -1000000.0;
   public static final double maxY = 1000000.0;
   public static final double minY = -1000000.0;
-  
-/*  public static final double maxX = 700.0;
+/*public static final double maxX = 700.0;
   public static final double minX = -100.0;
   public static final double maxY = 500.0;
   public static final double minY = -200.0;*/
 
-  // Copy constructor
-  // @param VoronoiDigram Base Voronoi diagram
-  // @return VoronoiDiagram Copy of base diagram
-  public VoronoiDiagram(
-    VoronoiDiagram other
-  ) {
-    edge = new ArrayList<HalfEdge>();
-    vertex = new ArrayList<Vertex>();
-    face = new ArrayList<Face>();
-    face_searcher = new HashMap<Point, Face>();
-    
-    // Hash map from others' diagram elements to new one
-    // If we already copied some element, we store this connection
-    // in these maps
-    HashMap<HalfEdge, HalfEdge> edges_map = new HashMap<HalfEdge, HalfEdge>();
-    HashMap<Vertex, Vertex> vertex_map = new HashMap<Vertex, Vertex>();
-    HashMap<Face, Face> faces_map = new HashMap<Face, Face>();
-   
-    for (HalfEdge e : other.getEdges()) {
-      edge.add(
-        deep_edge_copy(
-          e,
-          edges_map,
-          vertex_map,
-          faces_map
-        )
-      );
-    }
-    
-    for (Vertex v : other.getVertexes()) {
-      vertex.add(
-        deep_vertex_copy(
-          v,
-          edges_map,
-          vertex_map,
-          faces_map
-        )
-      );
-    }
-    
-    for (Face f : other.getFaces()) {
-      face.add(
-        deep_face_copy(
-          f,
-          edges_map,
-          vertex_map,
-          faces_map
-        )
-      );
-    }
-    
-    for (Map.Entry<Point, Face> entry : other.getFS().entrySet()) {
-      Point key = entry.getKey();
-      Face value = entry.getValue();
-      face_searcher.put(
-        key, 
-        deep_face_copy(
-          value,
-          edges_map,
-          vertex_map,
-          faces_map
-        )
-      );
-    }
+  private static final Point top_right = new Point(maxX, maxY);
+  private static final Point top_left = new Point(minX, maxY);
+  private static final Point bottom_right = new Point(maxX, minY);
+  private static final Point bottom_left = new Point(minX, minY);
 
-  }
-
-  private Vertex deep_vertex_copy(
-    Vertex v, 
-    HashMap<HalfEdge, HalfEdge> edges_map,
-    HashMap<Vertex, Vertex> vertex_map,
-    HashMap<Face, Face> faces_map
-  ) {
-    if (vertex_map.containsKey(v)) {
-      return vertex_map.get(v);
-    } else {
-      Vertex new_v = new Vertex(v.getX(), v.getY(), v.isInfinite());
-      vertex_map.put(v, new_v);
-      
-      new_v.setIncidentEdge(
-          deep_edge_copy(
-            v.getIncidentEdge(),
-            edges_map,
-            vertex_map,
-            faces_map
-          )
-      );
-      return new_v;
-    }
-  }
-  
-  private Face deep_face_copy(
-    Face f, 
-    HashMap<HalfEdge, HalfEdge> edges_map,
-    HashMap<Vertex, Vertex> vertex_map,
-    HashMap<Face, Face> faces_map
-  ) {
-    if (faces_map.containsKey(f)) {
-      return faces_map.get(f);
-    } else {
-      Face new_f = new Face(f);
-      faces_map.put(f, new_f);
-      
-      new_f.setIncidentEdge(
-          deep_edge_copy(
-            f.getIncidentEdge(),
-            edges_map,
-            vertex_map,
-            faces_map
-          )
-      );
-      return new_f;
-    }
-  }
-  
-  private HalfEdge deep_edge_copy(
-    HalfEdge e, 
-    HashMap<HalfEdge, HalfEdge> edges_map,
-    HashMap<Vertex, Vertex> vertex_map,
-    HashMap<Face, Face> faces_map
-  ) {
-    if (edges_map.containsKey(e)) {
-      return edges_map.get(e);
-    } else {
-      HalfEdgeBuilder new_e_builder = new HalfEdgeBuilder(e.getName());
-      edges_map.put(e, new_e_builder.getHalfEdgeReference());
-      new_e_builder
-        .setOrigin(
-            deep_vertex_copy(
-              e.getOrigin(),
-              edges_map,
-              vertex_map,
-              faces_map
-            )
-        )
-        .setNextEdge(
-          deep_edge_copy(
-            e.getNextEdge(),
-            edges_map,
-            vertex_map,
-            faces_map
-          )
-        )
-        .setPreviousEdge(
-          deep_edge_copy(
-            e.getPreviousEdge(),
-            edges_map,
-            vertex_map,
-            faces_map
-          )
-        )
-        .setTwinEdge(
-          deep_edge_copy(
-            e.getTwinEdge(),
-            edges_map,
-            vertex_map,
-            faces_map
-          )
-        )
-        .setLeftIncidentFace(
-          deep_face_copy(
-            e.getLeftIncidentFace(),
-            edges_map,
-            vertex_map,
-            faces_map
-          )
-        );
-      try {
-        return new_e_builder.getHalfEdge();
-      } catch (Exception exc) {
-        Debug.log(exc.getMessage());
-        return null;
-      }
-    }
-  }
+  // Bounding segemnts of Voronoi Diagram
+  public static final Segment[] bounds = new Segment[] {
+    new Segment(top_left, top_right),
+    new Segment(top_right, bottom_right),
+    new Segment(bottom_right, bottom_left),
+    new Segment(bottom_left, top_left) 
+  };
   
   public VoronoiDiagram(
     ArrayList<HalfEdge> e, 
@@ -210,64 +42,45 @@ public class VoronoiDiagram implements VisualData{
     ArrayList<Face> f,
     HashMap<Point, Face> fs
   ) {
-    edge = e;
-    vertex = v; 
-    face = f; 
-    face_searcher = fs; 
+    super(e, v, f, fs);
+  }
+  
+  /**
+   * Copy constructor
+   * @param VoronoiDiagram other diagram
+   * @return VoronoiDiagram Copy of other diagram (deep).
+   */
+  public VoronoiDiagram(VoronoiDiagram other) {
+    super(other); // deep copy of DCEL 
   }
 
-  public ArrayList<HalfEdge> getEdges() {
-    return edge;
-  }
-  
-  public ArrayList<Vertex> getVertexes() {
-    return vertex;
-  }
-
-  public ArrayList<Face> getFaces() {
-    return face;
-  }
-  
-  public HashMap<Point, Face> getFS() {
-    return face_searcher;
-  }
-  
-  // Constructor for 1 point
+  /** 
+   * Constructor for 1 point
+   * @param Point
+   * @return VoronoiDiagram of one point
+   */
   public VoronoiDiagram(Point a) {
-    edge = new ArrayList<HalfEdge>();
-    vertex = new ArrayList<Vertex>();
-    face = new ArrayList<Face>();
-    face_searcher = new HashMap<Point, Face>();
-    
+    super();
     Face f = new Face();
     f.setPoint(a);
     face_searcher.put(a, f);
     face.add(f);
-    // TODO: check if this is enough for algortihm
   }
 
-  // Constructor for 2 points
+  /** 
+   * Constructor for 2 points
+   * @param Point 
+   * @param Point
+   * @return VoronoiDiagram of two points
+   */
   public VoronoiDiagram(Point a, Point b) {
-    edge = new ArrayList<HalfEdge>();
-    vertex = new ArrayList<Vertex>();
-    face = new ArrayList<Face>();
-    face_searcher = new HashMap<Point, Face>();
-    
-    Point top_right = new Point(maxX, maxY);
-    Point top_left = new Point(minX, maxY);
-    Point bottom_right = new Point(maxX, minY);
-    Point bottom_left = new Point(minX, minY);
+    super(); // initialize empty DCEL
 
-    Segment[] bounds = new Segment[] {
-      new Segment(top_left, top_right),
-      new Segment(top_right, bottom_right),
-      new Segment(bottom_right, bottom_left),
-      new Segment(bottom_left, top_left) 
-    };
-
+    // Find support segment and its' bisector
     Segment support_segment = new Segment(a, b);
     Line perp_bisector = support_segment.getPerpendicularBisector();
 
+    // Find intersection of bisector with bounds
     ArrayList<Point> intersection_points = new ArrayList<Point>();
     for (Segment s : bounds) {
       try {
@@ -280,6 +93,7 @@ public class VoronoiDiagram implements VisualData{
 
     intersection_points = Geometry.excludeSamePoints(intersection_points);
 
+    // We always expect to have 2 intersection points
     assert(intersection_points.size() == 2);
     
     Point p1 = intersection_points.get(0);
@@ -288,6 +102,7 @@ public class VoronoiDiagram implements VisualData{
     Debug.log("intersection_point" + p1.toString());
     Debug.log("intersection_point" + p2.toString());
 
+    // "sort" points to have p1.y > p2.y
     if (p1.getY() < p2.getY()) {
       Point temp = new Point(p1.getX(), p1.getY());
       p1.setX(p2.getX());
@@ -301,7 +116,7 @@ public class VoronoiDiagram implements VisualData{
     vertex.add(v1);
     vertex.add(v2);
 
-    // TODO: add DCEL initialization
+    // 3 faces
     Face f1 = new Face();
     Face f2 = new Face();
     Face f_inf = new Face();
@@ -314,14 +129,14 @@ public class VoronoiDiagram implements VisualData{
     face_searcher.put(a, f1);
     face_searcher.put(b, f2);
 
+    // Build half edges
     HalfEdgeBuilder e_builder = new HalfEdgeBuilder("e");
     HalfEdgeBuilder e_twin_builder = new HalfEdgeBuilder("e_twin");
     HalfEdgeBuilder e_inf_builder = new HalfEdgeBuilder("e_inf");
     HalfEdgeBuilder e_inf_twin_builder = new HalfEdgeBuilder("e_inf_twin");
     HalfEdgeBuilder e_twin_inf_builder = new HalfEdgeBuilder("e_twin_inf");
     HalfEdgeBuilder e_twin_inf_twin_builder = new HalfEdgeBuilder("e_twin_inf_twin");
-    //f1.setIncidentEdge(e_builder.getHalfEdgeReference());
-    //f2.setIncidentEdge(e_twin_builder.getHalfEdgeReference());
+    
     try {
       e_builder
         .setOrigin(v2)
@@ -371,31 +186,25 @@ public class VoronoiDiagram implements VisualData{
     }
   }
   
-  // constructor for 3 points
-  // TODO: ADD TRANGLE CASE!!!!!
+  /** 
+   * Constructor for 3 points
+   * @param Point 
+   * @param Point
+   * @param Point
+   * @return VoronoiDiagram of three points
+   */
   public VoronoiDiagram(Point a, Point b, Point c) {
-    edge = new ArrayList<HalfEdge>();
-    vertex = new ArrayList<Vertex>();
-    face = new ArrayList<Face>();
-    face_searcher = new HashMap<Point, Face>();
+    super(); // initialize empty DCEL structure
     
-    Point top_right = new Point(maxX, maxY);
-    Point top_left = new Point(minX, maxY);
-    Point bottom_right = new Point(maxX, minY);
-    Point bottom_left = new Point(minX, minY);
-
-    Segment[] bounds = new Segment[] {
-      new Segment(top_left, top_right),
-      new Segment(top_right, bottom_right),
-      new Segment(bottom_right, bottom_left),
-      new Segment(bottom_left, top_left) 
-    };
-
+    // Three support segments and their bisectors
     Segment AB = new Segment(a, b);
     Segment BC = new Segment(b, c);
     Segment CA = new Segment(c, a);
+
     Line AB_bisector = AB.getPerpendicularBisector();
     Line BC_bisector = BC.getPerpendicularBisector();
+    
+    // Find point of bisector intersection
     Point bisector_intersection;
     try {
       bisector_intersection = Geometry.intersect(AB_bisector, BC_bisector);
@@ -408,16 +217,20 @@ public class VoronoiDiagram implements VisualData{
     Point BC_middle = Geometry.getMiddlePoint(BC);
     Point CA_middle = Geometry.getMiddlePoint(CA);
 
+    // Rays from the center of triangle
     Ray AB_ray = new Ray(bisector_intersection, AB_middle);
     Ray BC_ray = new Ray(bisector_intersection, BC_middle);
     Ray CA_ray = new Ray(bisector_intersection, CA_middle);
 
-    //!!!!!!!!!!!! TODO: ADD TRANGLE CASE!!!!!
+    // Check that (a,b,c) listed in clockwise order.
+    // Use sign in other case
     double sign = 1.0;
     Line l = new Line(a, c);
     if (l.checkPointVerticalAlignment(b) == Line.POINT_BELOW_LINE) {
      sign = -1.0;
     }
+
+    // If triangle (a,b,c) has angle > Pi/2 than we need to change direction of some ray
     if (sign*Geometry.crossProduct(AB.getDirection(), new Vector(a, bisector_intersection)) > 0) {
       AB_ray = new Ray(AB_middle, bisector_intersection);
     }
@@ -470,11 +283,14 @@ public class VoronoiDiagram implements VisualData{
     Debug.log("intersection_point" + p1.toString());
     Debug.log("intersection_point" + p2.toString());
     Debug.log("intersection_point" + p3.toString());
-    
+   
+    // Initialize vertexes 
     Vertex center = new Vertex(bisector_intersection.getX(), bisector_intersection.getY(), false);
     Vertex v1;
     Vertex v2 = new Vertex(p2.getX(), p2.getY(), true);
     Vertex v3;
+
+    // If he have counter-clockwise order, swap vertexes
     if (sign  < 0)
       Debug.log("Invert points!");
     if (sign < 0) {
@@ -494,6 +310,7 @@ public class VoronoiDiagram implements VisualData{
     vertex.add(v2);
     vertex.add(v3);
 
+    // Initialize faces
     Face f1 = new Face();
     Face f2 = new Face();
     Face f3 = new Face();
@@ -519,7 +336,7 @@ public class VoronoiDiagram implements VisualData{
     face.add(f3);
     face.add(f_inf);
     
-    // TODO:
+    // Initialize half-edges
     ArrayList<HalfEdgeBuilder> builder = new ArrayList<HalfEdgeBuilder>();
     for (int i = 0; i < 13; i++) {
       Integer name = new Integer(i);
@@ -612,154 +429,15 @@ public class VoronoiDiagram implements VisualData{
     } catch (VoronoiHalfEdgeIsNotValidException e) {
       Debug.log(e.getMessage());
     }
-    check();
-  }
- 
-  // TODO: refactor this method
-  public Face getFaceAroundPoint(Point point) throws NoDataException {
-    //Debug.log("Search for " + point.toString() + "in " + face_searcher.toString());
-    for (Point p : face_searcher.keySet()) {
-      if (p.getX() == point.getX() && p.getY() == point.getY()) {
-        Face f = face_searcher.get(p); 
-        //Debug.log("face found" + f.toString());
-        return f;
-      }
-    }
     
-    throw new NoDataException();
-  }
-
-  public boolean check() {
-    Debug.log("checking result...");
-    HashSet<Face> used_face = new HashSet<Face>();
-    HashSet<Vertex> used_vertex = new HashSet<Vertex>();
-    HashSet<HalfEdge> used_edge = new HashSet<HalfEdge>();
-
-    String s = "";
-    for (Point p : face_searcher.keySet()) {
-      Debug.log("---------------------------------------");
-      s += "For point " + p + "\n\n";
-
-      Face f = face_searcher.get(p);
-      used_face.add(f);
-      s += "Found face: \n";
-      HalfEdge start = f.getIncidentEdge();
-      HalfEdge temp = start;
-      int iteration = 0;
-      do {
-        iteration++;
-        if (iteration > 100) {
-          s += "INFINITE CYCLE!\n\n";
-          Debug.log(s);
-          return false;
-        }
-        used_edge.add(temp);
-        Vertex v = temp.getOrigin();
-        used_vertex.add(v);
-        s += "face edge = "+ temp + "\n\n";
-        
-        if (temp.isInfinite()) {
-          s += "this is infinite edge. Start to show inf. face edges\n";
-          HalfEdge twin = temp.getTwinEdge();
-          HalfEdge e = twin;
-          int it = 0;
-          do {
-            it++;
-            if (it > 100) {
-              s += "INFINITE CYCLE!\n";
-              Debug.log(s);
-              return false;
-            }
-            used_edge.add(e);
-
-            s += "  face edge = "+ e + "\n";
-            e = e.getNextEdge();
-          } while (e != twin);
-          used_face.add(twin.getLeftIncidentFace());
-          s += "end of showing inf. face edges\n";
-        }
-        if (temp.getLeftIncidentFace() != f) {
-          s += "Incorrect face!! "+ temp.getLeftIncidentFace()+ "\n";
-              Debug.log(s);
-          return false;
-        }
-        temp = temp.getNextEdge();
-      } while(start != temp);
-
-    }
-    
-    Iterator<Vertex> it = vertex.iterator();
-    while(it.hasNext()) {
-      Vertex v = it.next();
-      if (!used_vertex.contains(v)) {
-        s+="Removing vertex" + v + "\n";
-        it.remove();       
-        v = null;
-      } else {
-        if (!used_edge.contains(v.getIncidentEdge())) {
-          s += "Wrong link from vertex " + v + "to edge " + v.getIncidentEdge() + "\n";
-              Debug.log(s);
-          return false;
-        }
-      }
-    }
-
-    Iterator<Face> fit = face.iterator();
-    while(fit.hasNext()) {
-      Face f = fit.next();
-      s += "Considering Face = " + f + "\n";
-      if (!used_face.contains(f)) {
-        s += "Removing face" + f;
-        fit.remove();       
-        f = null;
-      } else {
-        if (!used_edge.contains(f.getIncidentEdge())) {
-              Debug.log(s);
-          Debug.log("Wrong link from face " + f + "to edge " + f.getIncidentEdge());
-          return false;
-        }
-      }
-    }
-
-    Iterator<HalfEdge> hit = edge.iterator();
-    while(hit.hasNext()) {
-      HalfEdge e = hit.next();
-      if (!used_edge.contains(e)) {
-        s += "Removing edge" + e;
-        hit.remove();       
-        e = null;
-      } else {
-        if (!used_edge.contains(e.getNextEdge())) {
-          Debug.log(s);
-          Debug.log("Wrong link from edge.next " + e + "to edge " + e.getNextEdge());
-          return false;
-        }
-        if (!used_edge.contains(e.getPreviousEdge())) {
-          Debug.log(s);
-          Debug.log("Wrong link from edge.prev " + e + "to edge " + e.getPreviousEdge());
-          return false;
-        }
-        if (!used_edge.contains(e.getTwinEdge())) {
-          Debug.log(s);
-          Debug.log("Wrong link from edge.twin " + e + "to edge " + e.getTwinEdge());
-          return false;
-        }
-      }
-    }
-    
-    
-    Debug.log("checking was succesful");
-    Debug.log("Edges: " + edge.size());
-    Debug.log("Vertex: " + vertex.size());
-    Debug.log("Face: " + face.size());
-
-
-    Debug.log("---------------------------------------");
-    return true;
+    // Check we have correct DCEL structure
+    // and delete unused elements (all to right from L diagram
+    // and all to left from R)
+    checkAndClean();
   }
 
   public void render(Graphics2D g) {
-    Debug.log("rendering voronoi...");
+    Debug.log("Rendering voronoi...");
     for (HalfEdge e : edge) {
       if (!e.isInfinite()) {
         g.setColor(Color.red);
@@ -782,7 +460,7 @@ public class VoronoiDiagram implements VisualData{
 
   @Override
   public String toString() {
-    String s = "{ Voronoi Diagram Structure \nPoints = [";
+    String s = "{ Voronoi Diagram \nPoints = [";
     for (Point v : face_searcher.keySet()) {
       s += "(" + v.getX() + ", " + v.getY() + "), ";
     }
